move enums to deprecated
