+++ packages/components/src/index.ts
@@ -7,7 +7,5 @@ export { default as Meta } from './components/Meta'
 export { default as SVG } from './components/SVG'
 export * from './components/viewers/index'
 export type * from './components/viewers/types'
-export { FieldState, FieldType, FormResultState } from './lib/enums'
-export type { FieldStateValue, FieldTypeValue, FormResultStateValue } from './lib/enums'
 export { SourceComponents, SourceProvider } from './lib/provider'
 export { theme } from './theme'
